fixup! feat!: integrate classes from FlexibleMovement
