+++ src/main/java/org/terasology/module/behaviors/actions/MoveToAction.java
@@ -17,10 +17,6 @@ import org.terasology.engine.registry.CoreRegistry;
 import org.terasology.engine.registry.In;
 import org.terasology.engine.world.WorldProvider;
 import org.terasology.engine.world.block.Blocks;
-import org.terasology.flexiblemovement.components.MinionMoveComponent;
-import org.terasology.flexiblemovement.plugin.MovementPlugin;
-import org.terasology.flexiblemovement.plugin.WalkingMovementPlugin;
-import org.terasology.flexiblemovement.systems.PluginSystem;
 import org.terasology.module.behaviors.components.MinionMoveComponent;
 import org.terasology.module.behaviors.plugin.MovementPlugin;
 import org.terasology.module.behaviors.plugin.WalkingMovementPlugin;
